+++ Source/Libraries/CorruptCore/SaveStateKey.cs
@@ -1,0 +1,34 @@
+namespace RTCV.CorruptCore
+{
+    using System;
+    using System.Collections.Generic;
+    using Ceras;
+
+    [Serializable]
+    [Ceras.MemberConfig(TargetMember.All)]
+    public class SaveStateKey
+    {
+        public StashKey StashKey = null;
+        public string Text = "";
+
+        public SaveStateKey(StashKey stashKey, string text)
+        {
+            StashKey = stashKey;
+            Text = text;
+        }
+    }
+
+    [Serializable]
+    [Ceras.MemberConfig(TargetMember.All)]
+    public class SaveStateKeys
+    {
+        public string VanguardImplementation { get; set; }
+        public List<StashKey> StashKeys = new List<StashKey>();
+        public List<string> Text = new List<string>();
+
+        public SaveStateKeys()
+        {
+            VanguardImplementation = (string)RTCV.NetCore.AllSpec.VanguardSpec?[VSPEC.NAME] ?? "ERROR";
+        }
+    }
+}
